userauth: _libssh2_userauth_publickey() pubkeydata_len -> size_t
Matching its public-facing wrapper and internal callers.

+++ src/userauth.c
@@ -1415,7 +1415,7 @@ _libssh2_userauth_publickey(LIBSSH2_SESSION *session,
                             const char *username,
                             unsigned int username_len,
                             const unsigned char *pubkeydata,
-                            unsigned long pubkeydata_len,
+                            size_t pubkeydata_len,
                             LIBSSH2_USERAUTH_PUBLICKEY_SIGN_FUNC
                             ((*sign_callback)),
                             void *abstract)
